backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
bcrypt==4.0.1
boto3==1.42.83
botocore==1.42.83
click==8.3.2
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
idna==3.11
jmespath==1.1.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.43.0
